fix search error
This commit is fixing the error where the search bar was throing errors on every new character typed. It did this because the some function we were using threw errors if a tag did not exist at all on an object. Another fix would have been to initialize every object with an empty list of tags, but I believe this solution is more robust for now since it avoids using memory needlessly. I will figure out whether or not I will require tags later.